requirements.txt
aniso8601==9.0.1
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.1.0
Flask==2.2.2
Flask-RESTful==0.3.9
idna==3.4
importlib-metadata==6.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==22.0
pluggy==1.0.0
pytest==7.2.0
pytest-envfiles==0.1.0
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
